FIX: Existing User detection

+++ src/components/SignUp.jsx
@@ -21,34 +21,45 @@ const Signup = () => {
     const username = event.target.name.value;
     const email = event.target.email.value;
     const password = event.target.password.value;
-
+  
     if (!email || !password || !username) {
-      setError('Please fill in all fields.');
+      setError('Please fill in all the fields.');
+      console.log("fill in all the fields.");
       return;
     }
-
+  
     const existingUsers = JSON.parse(localStorage.getItem('users')) || [];
-
+  
+    // Check if a user with the same email or username already exists
+    const userExists = existingUsers.some((user) => user.email === email || user.username === username);
+  
+    if (userExists) {
+      setError('An account with the same email or username already exists.');
+      console.log("account with the same email or username already exists.");
+      return;
+    }
+  
     // Add the new user to the existing users array
     existingUsers.push({ username, email, password });
     // Save the updated users array back to local storage
-    localStorage.setItem('users', JSON.stringify(existingUsers)); 
-
-    // Dispatch an action to update the global state with the new user1 data
-    dispatch(
-      {
-        type: "SIGNUP_USER",
-        payload: { username, email }
-      });
+    localStorage.setItem('users', JSON.stringify(existingUsers));
+  
+    // Dispatch an action to update the global state with the new user data
+    dispatch({
+      type: "SIGNUP_USER",
+      payload: { username, email }
+    });
+  
     setSubmitted(true);
     setError('');
   };
 
 
   if (submitted) {
+    console.log("signed up successfully");
     return (
       <div className="container">
-        <h4 className='text-center text-success mb-3'>Signup successful! You can now login.</h4>
+        <h4 className='alert alert-success text-center text-capitalize'>Yay! Sign In Now.</h4>
           <SignIn />
       </div>
     );
@@ -57,6 +68,7 @@ const Signup = () => {
   else {
     return (
       <div className="container" id='form'>
+        {error && <h4 className='alert alert-danger text-center text-capitalize'>{error}</h4>}
         <h2>Sign Up</h2>
         <h6>To avail exclusive offers</h6>
         <form onSubmit={handleSignup}>
@@ -72,7 +84,6 @@ const Signup = () => {
             <label htmlFor="password">Password</label>
             <input type="password" className="form-control" id="password" />
           </div>
-          {error && <p className="text-danger">{error}</p>}
           <button type="submit" className="custom-button mt-4 mb-3 float-end">Sign Up</button>
         </form>
       </div>
